new AskForGameStart in GrapicInterface, Game.Save have a new type of saving by user, program has new While

+++ beadandoV3/Ship.cs
@@ -49,7 +49,7 @@ namespace beadandoV3
 
         public override string ToString()
         {
-            return player + "," + point.x + "," + point.y + "," + size + "," + orientation;
+            return (int)player + "," + point.x + "," + point.y + "," + size + "," + (int)orientation;
         }
     }
 }
